initial commit, move logging setup teardown to server
